# Implement `vec_slice<-` in C