--- a/R/slice.R
+++ b/R/slice.R
@@ -43,11 +43,19 @@
 #' x <- sample(10)
 #' x
 #' vec_slice(x, 1:3)
-#' vec_slice(x, 2L) <- 100
+#'
+#' # You can assign with the infix variant:
+#' vec_slice(x, 2) <- 100
 #' x
 #'
-#' vec_slice(mtcars, 1:3)
+#' # Or with the regular variant that doesn't modify the original input:
+#' y <- vec_assign(x, 3, 500)
+#' y
+#' x
 #'
+#'
+#' # Slicing objects of higher dimension:
+#' vec_slice(mtcars, 1:3)
 #'
 #' # Type stability --------------------------------------------------
 #'
@@ -95,11 +103,13 @@ vec_slice_fallback <- function(x, i) {
   vec_restore(out, x)
 }
 
-#' @export
 #' @rdname vec_slice
+#' @export
 `vec_slice<-` <- function(x, i, value) {
   .Call(vctrs_assign, x, i, value)
 }
+#' @rdname vec_slice
+#' @export
 vec_assign <- function(x, i, value) {
   .Call(vctrs_assign, x, i, value)
 }