--- a/src/cast.c
+++ b/src/cast.c
@@ -469,6 +469,31 @@ static SEXP vec_restore_dispatch(SEXP x, SEXP to, SEXP i) {
                          syms_i, i);
 }
 
+// [[ include("vctrs.h") ]]
+SEXP vec_coercible_cast(SEXP x, SEXP to, struct vctrs_arg* x_arg, struct vctrs_arg* to_arg) {
+  // Called for the side effect of generating an error if there is no
+  // common type
+  int _left;
+  vec_type2(x, to, x_arg, to_arg, &_left);
+
+  return vec_cast(x, to);
+}
+
+// [[ register() ]]
+SEXP vctrs_coercible_cast(SEXP x, SEXP to, SEXP x_arg_, SEXP to_arg_) {
+  if (!r_is_string(x_arg_)) {
+    Rf_errorcall(R_NilValue, "`x_arg` must be a string");
+  }
+  if (!r_is_string(to_arg_)) {
+    Rf_errorcall(R_NilValue, "`to_arg` must be a string");
+  }
+
+  struct vctrs_arg_wrapper x_arg = new_wrapper_arg(NULL, r_chr_get_c_string(x_arg_, 0));
+  struct vctrs_arg_wrapper to_arg = new_wrapper_arg(NULL, r_chr_get_c_string(to_arg_, 0));
+
+  return vec_coercible_cast(x, to, (struct vctrs_arg*) &x_arg, (struct vctrs_arg*) &to_arg);
+}
+
 
 void vctrs_init_cast(SEXP ns) {
   syms_vec_cast_dispatch = Rf_install("vec_cast_dispatch");