… NAMESPACE
… NEWS.md
… src/slice.c
… src/utils.c
… src/utils.h
… src/vctrs.h
… tests/testthat/helper-s3.R
… tests/testthat/test-cast.R
… tests/testthat/test-slice-assign.R
… tests/testthat/test-slice.R