And add a `vec_assign()` variant that doesn't modify the input in the current environment.

Atomic assignment:

```r
bench::mark(
  before = vec_assign_before(letters, 1:15, ""),
  after = vec_assign(letters, 1:15, ""),
  base = base_assign(letters, 1:15, "")
)[1:6]
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 before      13.36µs  16.81µs    55647.      256B     11.1
#> 2 after        2.42µs   2.71µs   321932.      256B      0
#> 3 base         1.56µs   1.92µs   486901.      256B     48.7
```

Data frame assignment:

```r
df <- mtcars
val <- mtcars[1, ] * rnorm(ncol(mtcars))

bench::mark(
  before = vec_assign_before(df, 1:15, val),
  after = vec_assign(df, 1:15, val),
  base = base_row_assign(df, 1:15, val),
  check = FALSE # row.names currently not assigned
)[1:6]
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 before      103.5µs  115.1µs     8246.    3.27KB     15.1
#> 2 after        24.1µs     28µs    34065.    3.27KB     10.2
#> 3 base         71.7µs   81.7µs    11461.    3.27KB     15.2
```

Nested data frame assignment:

```r
nested <- tibble(nested = mtcars, !!!mtcars)
nested_val <- tibble(nested = val, !!!val)

bench::mark(
  before = vec_assign_before(nested, 1, nested_val),
  after = vec_assign(nested, 1, nested_val),
  tibble = base_row_assign(nested, 1, nested_val),
  check = FALSE
)[1:6]
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 before      194.9µs  212.4µs     4465.    6.53KB     12.8
#> 2 after        47.5µs   52.4µs    18084.    6.53KB     13.3
#> 3 tibble      115.7µs    128µs     7492.    6.53KB     15.4
```

Before the C implementation of cast and type2 for data frames, the above was much slower:

```r
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 before        906µs   1.08ms      883.    6.53KB     15.4
#> 2 after         738µs 842.69µs     1113.    6.53KB     15.3
#> 3 tibble        117µs 131.73µs     6848.    6.53KB     17.9
```

In general it seems essential for performance to implement cast and type2 in C to avoid dispatch. For instance, here are the results for factors, which are still implemented in R:

```r
x <- factor(c("foo", "bar", "foo", "foo"))
y <- factor(c("bar", "bar"))
i <- 1:2

bench::mark(
  before = vec_assign_before(x, i, y),
  after = vec_assign(x, i, y),
  base = base_assign(x, i, y)
)[1:6]
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 before      124.3µs  138.5µs     6823.        0B     14.8
#> 2 after         142µs  153.4µs     6165.        0B     14.8
#> 3 base         10.5µs   12.5µs    73867.        0B     14.8
```