@lionel- 2019-05-17
@hadley Now rebased.